ModuleTracker backed by memory [0] make get_source_paths a record field
Summary:
In order to allow unsaved changes, we need a way to inject in-memory source code
into the ModuleTracker so thay it can mask the filesystem when working with an
overlaid environment.

My preference is to handle this in the ReadOnly type by using a closure (as we
tend to do for ReadOnly environments), so this is a preperatory refactor where I
move the definition of the function out of ReadOnly module scope and into the
constructor, so that it becomes just another record field.

Reviewed By: grievejia

Differential Revision: D36220554

fbshipit-source-id: c41387d4939221c2c3f96ee8c202322f1f346224