Move callable fetching into its own file in the interprocedural directory
Summary:
When performing a taint analysis, we walk through the source code to find all callables (functions and methods) to analyze after performing a type check.
We want to improve that step and return a better data structure to represent its result. Right now, it returns a list of callables with a boolean indicating whether the callable is internal or not. We also return a `filtered_callables` set which is not really documented.

As a first step, this diff moves all the callable-fetching logic in its own module under the interprocedural directory.

Reviewed By: dkgi

Differential Revision: D36463520

fbshipit-source-id: 60eb33299dea4ec2f0f81586f522c8a5561c309f