Implement the intented behavior of analyzing __setitem__
Summary:
This diff implements the intented behavior of analzying `__setitem__` (as described in the previous diff).
We use the hardcoded model of `__setitem__` for any subtype of dict or unresolved callees: `base[index] = value`.
This is incorrect, but can lead to higher SNR, because we assume in most cases, we run into an expression whose type is exactly `dict`, rather than a (strict) subtype of `dict` that overrides `__setitem__`.

We also treat `e.__setitem__(k, v)` as `e = e.__setitem__(k, v)` where method `__setitem__` returns the updated self.
Due to modeling `__setitem__` with the assignment, the user-provided models of `__setitem__` will be ignored, if they are inconsistent with treating `__setitem__` as returning an updated self.

Reviewed By: arthaud

Differential Revision: D36415070

fbshipit-source-id: 83fe21144f251a23aed8ead6609b7d685a55ffae