s/ignore/handle_ignores_and_fixmes
Summary:
I was skimming the preprocessing code (as part of building
up context for how to have overlaid server state) and was
confused by the `ignore` stage of preprocessing, which sounds
like it's "ignoring" errors but is actually doing more than
that because it's flagging unused ignores/fixmes as well.

I figured this name is a bit clearer. It's not a big deal since
this is a private function.

Reviewed By: grievejia

Differential Revision: D36490597

fbshipit-source-id: f7af0119966503e418bd5beecec6c45a66aa9a11