Run pyre statistics on the project root if no directories given
Summary:
We need to provide basic coverage statistics support for distillery, meaning this pyre statistics script will be run in a project without local configurations.

Currently, if a local configuration is passed in with no filter paths, the entire local project is checked. Filter paths mean only sources within the provided directories are checked. If neither are passed in, the cwd is used.

This is somewhat confusing for projects without local configs -- we should make the explicit directory argument better named and documented, and then default to checking the entire project (regardless of whether it is defined with a local or global pyre config) when no explicit directories are provided.

Reviewed By: stroxler

Differential Revision: D36488621

fbshipit-source-id: edf804e980672de7797dc84af96c4594131d922b