Add some more explanation to `get_or_recompute_local_annotations`
Summary:
I was confused by this function at first since we *do* save local annotations
in `TypeEnvironment.LocalAnnotations`, but the doc comment says we don't.

The explanation, it turns out, is that the existing comment is half-true. We
*can* save local annotations but we choose not to except in pysa runs, and as
a result for LSP purposes we have to recompute them; this behavior is controlled
via flags set on the python side.

This is the sort of confusing edge case that would actually be great to have in a high-level
summary of how our server state works, but at the moment we have zero architectural
docs and an inline comment is better than nothing.

Reviewed By: grievejia

Differential Revision: D36483625

fbshipit-source-id: f8bf294413ceef39de67cd71bc703c5a9629dadc