check-unsaved: refactor the get_raw_code handling
Summary:
In order to support using in-memory code for testing, I wrote
a `make_get_raw_code` function but I made it too specific: it
hard-codes that we look up code and fall back to filesystem
reads.

For overlays to be clean, it's best to instead have the function
take the fallback logic (at the moment reading from filesystem)
as an input and wrap that with the in-memory reads.

This will allow us to create a module tracker from a parent by
wrapping the `get_raw_code` function from the parent with the
overlay intercepting calls, in a way that doesn't assume the parent
is always filesystem-backed.

Reviewed By: shannonzhu, grievejia

Differential Revision: D36481096

fbshipit-source-id: fdd03704f0679488f40b8ab1471b6872a132c433