unsaved-changes: make type environment tables first-class
Summary:
Just as I made all of the other environment tables first-class
in order to prepare the interfaces to allow overlayable memory,
so we also need to make the two TypeEnvironment tables first-class.

This is the last of the tables that we have to migrate. At some point
we may want to make the callgraph overlayable, but I don't think that
is needed for an MVP; we can just always use NullBuilder to avoid
modifying the callgraph when inside of overlays for now and come back
to it later if needed.

NOTE: in reality, we have little-documented flags that cause us to
normally not save LocalAnnotations. As a result, it would be possible, but
very confusing, to not make these overlayable since we only use them for
pysa. My view is that it would be inappropriate to do this even though it
would work okay right now; we ought to make the entire TypeEnvironment
overlayable and not just part of it.

Reviewed By: grievejia

Differential Revision: D36482943

fbshipit-source-id: 9c3dd329946fd6b29c6482f0383d5b2f0387896d