Isolate preprocessing dependencies
Summary:
Something that has made it harder for me to reason about the
behavior of the bottom three layers of the environment stack
(module tracker + ast env + unannotated global env) is that
these aren't environment-table-backed envs so all the reasoning
about them has to be very special case.

What's been particularly vexing for me is that in theory any
environment could try to `get_raw_source` and trigger an arbitrary
dependency between `AstEnvironment.RawSources` and some other unit
of analysis.

This doesn't actually happen - RawSources is used *exclusively*
to produce preprocessed sources that populate UnannotatedGlobalEnvironment,
the only other use of raw sources is in error post processing (which
does not require dependency tracking) where we handle ignores and
syntax errors.

In this commit I push that knowledge into the type system by
- creating a new dependency key, private to AstEnvironment,
  representing cross-qualifier preprocessing dependencies
- removing `triggered_dependencies` from AstEnvironment.UpdateResult.
  These dependencies were never used, they were pure overhead so this
  is a (small) perf win in addition to being a code simplification.

The resulting code is quite a bit easier to reason about for some use cases because now the fanout from an AST change has been made explicit: it can only be module-level dependencies, they can only be introduced via preprocessing, and they are entirely captured by the `invalidated_modules` passed as input to `UnannotatedGlobalEnvironment`.

As a result, for example, it is now very clear that when propagating parent environment updates to an overlay we can skip `AstEnvironment` entirely and start the propagation with `UnannotatedGlobalEnvironment`.

Reviewed By: grievejia

Differential Revision: D36518448

fbshipit-source-id: a97cb5b065167010fd7e3e173d2fd93e33fa28b3