Simplify the DependencyKey logic
Summary:
We over-generalized `DependencyTrackedMemory` by creating a
fairly complex `In` module type that has to be painstakingly
created in `SharedMemoryKeys`.

This has several downsides:
- it puts a lot of logic that in my opinion belongs in
  DependencyKey.Make in a separate module
- we wind up writing quite a bit of extra code that isn't
  necessary (notice that there's a more red than green in
  this diff!)
- it couples that logic tightly to the `SharedMemoryKeys.dependency`
  type... this is in my way because I want to stop using that
  dependency key for `RawSources` (it obscures the control
  flow and comes with a silly efficiency hit)

The only benefit is that the test for DependencyTrackedMemory
uses a lighter-weight implementation that doesn't have a Registry,
which might be handy if that performance cost of that one test
really mattered but it doesn't.

So, I'm moving the definition of the Registry into
DependencyTrackedMemory. This makes the underlying `Key`
module type trivial instead of very complex, which simplifies
the test code and will enable us to easily separate out some
dependency keys from others if we would like to.

A bonus is that instead of having lots of nitty-gritty
plumbing code related to regestration and whatnot, the
SharedMemoryKeys module becomes very simple and doesn't contain much
besides a declaration of what the key types are - this is great,
since having that in easy-to-skim form really helps with
understanding how environments work :)

Reviewed By: grievejia

Differential Revision: D36502347

fbshipit-source-id: dbb75b1e9cdcf5bb1c8024801130ac82472a7b48