Consistently use the same definition for the taint analysis and call graph building
Summary:
When performing the taint analysis, we first gather all callables with the `FetchCallables` API.
We then compute call graphs of callables by walking other sources with `Preprocessing.defines`.
Finally, we compute the fixpoint using the callables gathered with `FetchCallables`.

This inconsistency leads to bugs where we are analyzing a function without having its call graph. This happens for instance when there are multiple functions with the same name.

This diff fixes the inconsistency by using the set of callables from `FetchCallables` when building the call graph.

Differential Revision: D36502567

fbshipit-source-id: e249794fa256a28829b751cb2a53e4ee9d3f164c