Move the override graph logic in its own file
Summary:
The logic that builds the override graph currently lives in the `DependencyGraph` modules.
This makes it pretty confusing as this module mixes up the override graph, call graph and dependency graph.
To make this cleaner, this diff moves the override graph logic in its own file.
This also cleans up the interface, by creating two submodules `Heap` and `SharedMemory` for the two different representations.
Finally, we now store the override graph directly in shared memory when using caching, instead of going through s-expressions.

Differential Revision: D36499220

fbshipit-source-id: 8b7833d2ac2cd382d29c588f1dee0183f70a924e