Use a new representation for the initial callables
Summary:
See the previous diff for context: D36463520 (https://github.com/facebook/pyre-check/commit/0af38a29231b07dfa4295f4b19860a6095032b3f)

This diff implements a new representation for the set of callables, removing the boolean and adding documentation.

Most importantly, this implements `Target.get_definitions` which is now the source of truth for the mapping of callable to definition.
This will be used to fix inconsistencies between the call graph and the taint analysis (see follow-up).

This also shows explicitly in the implementation an edge case that was initially ignored: having multiple definitions with the same name. We now raise a proper warning.

This adds proper `FetchCallables.get_xxx` APIs.

Finally, this removes the concept of `filtered` callables. Callables in unit tests are ignored directly when collecting callables. This simplifies the code.

Differential Revision: D36467465

fbshipit-source-id: 3750cdabb179f82c55e8b2bcd8b8b9f253acec75