Print the return code of the Worker_exited_abnormally exception
Summary:
# Problem
Right now, if a worker crashes, we get the following message: `Pyre encountered an internal exception: (exn.ml.Finally ("Worker.Worker_exited_abnormally(3297, _)")`. This is not really helpful since we do not know why the worker failed.

# Solution
Let's add logic to print the return code or signal that caused the crash.

Reviewed By: grievejia

Differential Revision: D36529430

fbshipit-source-id: 3cd1764d5dd32bb7d1ceacc098914e9b59574d66