Fix backward names in test
Summary:
I'm about to work on a dommit that adds a semantic type
`PyrePath.Analysis.t` representing a "built" source path.

One of the tests actually mixed up the semantics, and was
naming the input of `lookup_source` as `test_source_path`
which is backward - `lookup_source` takes an artifact
path as input!

Reviewed By: shannonzhu

Differential Revision: D36567487

fbshipit-source-id: d433350d3714fbf901e06b79108e86abd9e64f54