Type safe paths [0] Add PyrePath.Built
Summary:
**Stack Summary**

In this stack, I will create semantic types for pyre paths that
have different meanings.

The motivation is to make the interactions between BuildSystem
and the analysis codebase type-safe:
- because a build system like buck can remap source paths to
  different locations inside a project, we have to perform our
  analysis on "post-build" paths which means that the paths we
  pass to a ModuleTracker need to be output by the build system.
  - This isn't important in most tests because tests (and the
    standard OSS setup) have no-action build systems
  - But as a result, it would be very easy to break this and
    not notice because all unit tests might pass... the fact
    that we often can get away with skipping this is all the more
    reason to encode it in the type system!
- Moreover, because only the original paths are meaningful to users,
  we always have to map paths back at the end of analysis. This
  problem affects both errors (where it's handled by `instantiate`)
  and pyre infer results.

In an effort to encode this in the type system, I'm adding a
newtype PyrePath.Built.t, which will always indicate a path that
has been mapped to the analysis directory structure (meaning that it
has been handled by the build system if necessary).

Under the hood the newtype is exactly the same as PyrePath.t
but the type is abstract so ocaml will enforce correctness. Using them
will benefit us in two ways:
- we'll get more type safety - the `analysis` directory will deal
  *exclusively* with `Path.Built.t` which will make it much harder
  to accidentally forget to use the BuildSystem to convert paths.
- it will also make the types of functions more descriptive... today
  BuildSystem.update, for example, maps a `PyrePath.t list` to a
  `PyrePath.t list` but after this stack it will map a
  `PyrePath.t list` to a `PyrePath.Built.t list` which is
  much more descriptive.

**Diff Summary**

This diff creates the newtype, which is as yet unused.

To make that work I had to also create a module `PyrePath.Raw` so that
I could refer to the type as `Raw.t` (ocaml won't let you have a line like
`type t = t` when there's a `t` in the surrounding module scope, even though
it *does* allow that with `let` bindings)

Reviewed By: pradeep90

Differential Revision: D36563570

fbshipit-source-id: c39a9bebf9e04bc123ef3321d27fe0af8d3d0933