cold_start: Add project_qualifiers to APIs
Summary:
I'm working on a stack that will replace the existing `update`
logic for environments, which takes a `trigger` that can be `ColdStart`,
and instead separating cold start vs update into two separate functions.

This will have several benefits
- get us a ~5% perf gain in incremental cold start, by eliminating forks
- make the control flow easier to reason about
- greatly simplify the API refactors needed for overlays (why? Because we
  don't need cold start for overlays, and it's very hard to capture this
  in the type system given the current setup).

This is the second commit, where I add an API to get all the modules that
are in the project. I need this because in order to stop using `update`, we
need some reasonable way to get the module list for type checking; the truth
is that this module list does not depend on update logic *at all* in the cold
start case, it comes straight from ModuleTracker.

Reviewed By: grievejia

Differential Revision: D36615012

fbshipit-source-id: 3b527067d2b138b4c409517e488a3375d6faef8e