Split up `modelTest.ml` to avoid tests clobbering one another
Summary:
This is a temporary change: I discovered when refactoring the environment
stack to separate out cold_start vs update that it caused some pysa tests
to clobber mutable global state.

This is because of bugs in our environment stack handling where environments
share state and aren't reliably created vs updated using the correct APIs (these
bugs have gone undetected because under the hood all environments shared state,
but that needs to stop being true).

Unfortunately in order to track down and fix the bugs I needed to go through a series
of refactors, and it is neither worth the time nor illuminating to restack the changes. As
a result, my stack includes:
- this diff, where I split up and reorder `modelTest` to unblock my `cold_start` change
- after shipping that diff and fixing some other bugs, I back out this diff in D36640769

I chatted with arthaud about this and he's cool with it

Reviewed By: arthaud

Differential Revision: D36632156

fbshipit-source-id: 4154782432364cabdb77ff0a8f6aefb6816eff86