cold_start [0] Rename AstEnvironment update result
Summary:
In this stack, my intent is to separate out cold-start logic from
update logic. This will
(a) make our control flow clearer, since the two operations are in
    fact pretty different (especially now that we have lazy parsing,
    cold_start is very nearly a no-op!)
(b) provide around a 5% speed increase in incremental cold start,
    because we'll stop doing an unnecessary `fork` in order to
    map-reduce updates over empty triggered_dependencies which
    is what happens today... we don't have direct speed measurements
    on incremental cold start bu I know the gain is 5% because that's
    what we currently *lose* if I cause `check` to perform a
    `FineGrained` update.
(c) make it much easier to define a sensible API for overlays, which
    don't require a cold start at all but rather just an update

This is the first diff of the stack - I strip down
`AstEnvironment.UpdateResult` to just it's contents and change the
name to `AstEnvironment.InvalidatedModules` because this type needs to
be the output of both update *and* cold_start, and therefore needs a
better name.

Note that every single callsite got simpler when I did this, which is
a signal that this is a good simplification :)

Reviewed By: grievejia

Differential Revision: D36610396

fbshipit-source-id: 8c7178894b576bc998de6c0e53f4a3c39bee6339