Set the managed cache behavior in a sane way
Summary:
For reasons of both incremental performance and code modularity that
will make working with unsaved-change-overlays easier, I'm about to split
out the cold-start logic to be separate from update logic.

This will work much better, but UnmanagedCache has a hidden side-effect (which
isn't visible in behavior, but does have implications for runtime performance)
where `update_this_and_all_preceding_environments` sets the flag for using
shared memory or using a per-proces "unmanaged cache".

As a result, it's important to set this flag at some different point; my view
is that `create` is the right time to do it since this flag is never safe
to change after creation.

Moreover, since we have (in D36578363) made it impossible to update a
non-fine-grained environment this means we can get rid of the `clear` logic,
and therefore `update_this_and_all_preceding_environments` doesn't need to be
overridden (we can just use the normal EnvironmentTable implementation).

Reviewed By: grievejia

Differential Revision: D36632382

fbshipit-source-id: 8cb65c3abe2b5857d82b575d02ca559a7b4f1719