Stop re-using environments in tests
Summary:
When I initially rolled out using first-class tables for environments, I was worried that the
tests (which often create environments over and over) would eat up too much memory if I didn't
force `create` to reuse the same first-class table.

Now that we have lazy parsing, I'm not very worried about this anymore - the test environments
are tiny and each test module runs in an independent ocaml process, so the RAM use isn't a big
concern.

This commit flips us to not reusing environments. In the process, I discovered that multiple tests
were inappropriately using `create` + `update` operations to create a new environment stack and
then update it, implicitly assuming that the new environment would share all data with the old
one(!). This is not desirable, and I fixed all such tests.

Reviewed By: grievejia

Differential Revision: D36640770

fbshipit-source-id: a5f741f2f68b93ddd08925090596179c4e7319f5