cold start: Add a new cold_start API
Summary:
In order to keep the core logic isolated in an easy-to-review way
I've left the update functionality unchanged in this commit, but
I'm adding a new `cold_start` function to the entire environment stack.

In the next commit I'll remove the `ColdStart` trigger type and flip all
callsites to using `cold_start` instead.

I had hoped to strip out some of the trigger logic, but it turns out that
I cannot do that yet because as of this diff, all newly-created environments
still share data - the reason is that in order to avoid tests consuming too much
RAM prior to lazy parsing, I made base environments always use the same
tables. As a result, the `cold_start` function is currently doing almost the
same work that `update` does, and has to return an `UpdateResult` so that
subsequent environments can do their thing.

This is till an improvement though. Separating `cold_start` creates a
much cleaner API that expresses our intent clearly at each callsite
and unblocks further refactors in later commits.

Later in this stack (D36640770) I fix this (which is fine because of lazy
parsing) and I'll simplify the logic on top of that commit.

Reviewed By: grievejia

Differential Revision: D36615003

fbshipit-source-id: c9733bb615fbc8d79edf025c81a7eaad64c3490a