Use consistent names in UnannotatedGlobalEnvironmentTest
Summary:
Let's alwasy call the ReadOnly.t value `read_only`.

In one test it was being called `unannotated_global_environment`, which
is annoying for the inconsistency but not so bad; in another test it was
being called `ast_environment`, which is incredibly confusing and pretty
clearly just an oops from the past.

Reviewed By: pradeep90

Differential Revision: D36632381

fbshipit-source-id: 982b3daa247fa5b19d66a617f8808bf8c8a4b004