Add constraint to match on parent class decorators
Summary:
This refactors the existing `DecoratorConstraint` into its own module to prevent name clashes, and so it can be reused in both the `Decorator(...)` context and the `parent.decorator(...)` context. This is the first part in adding support for a new constraint that allows querying on class decorator.

For example, we may want to match on attributes or methods on only classes decorated with a certain decorator:
```
class C:
   ...

evil
class Foo(C):
   def f(self, a):
        // we want to taint a

class Bar(C):
    def f(self, a):
        // don't taint a
```

Reviewed By: arthaud

Differential Revision: D36302401

fbshipit-source-id: 4bcc4ec32e88d6f2c9268060b7566139b83e3fb6