Change PyrePath.Built to ArtifactPath
Summary:
Instead of using a tagged path type PyrePath.Built that behaves like a
path but is a distinct type to indicate analysis paths (which are controlled
by the build system: watchman updates always have to be converted *to* this type
and error messages have to be converted *away* from this type before they
are useful), instead use a wrapper type ArtifactPath.

The resulting code a tiny bit more verbose because we can't do things like
directly create a relative ArtifactPath or directly get the absolute path
as a string, but the separation of concerns between PyrePath and
ArtifactPath is clearer.

One operation we need *constantly* for tests is to create an artifact relative
to a root, which is a function I removed. To keep the code simple I added a
helper function `Test.relative_artifact_path` so that we don't have to chain
function calls to do this, but we don't expose that logic in production either.

NOTE: there's a separate question of whether to also have a semantic
type for source paths. I put that in a separate commit because it's easier to read
the changes this way.

Reviewed By: grievejia

Differential Revision: D36665185

fbshipit-source-id: a4292f6f5f1d1acc2780769148eb8b02881c28d9