Rename SourcePath -> ModulePath
Summary:
It was suggested in a late review of D36563570 (https://github.com/facebook/pyre-check/commit/e278277a31b05db11d05bb0e401524457f0a686b) that it would be
good to have semantic types for both "source path" and "artifact path".

But the fact that we have a type called `Ast.SourcePath.t` would make that
very confusing today.

Let's rename `SourcePath` to `ModulePath`, which is a better description
anyway - the type represents a "logical path" that can be used to locate
a module by name (i.e. by qualifier) relative to `SearchPaths.t` rather than
an actual directy-represented path, and I think `ModulePath` is more suggestive
of this kind of thing. Also, the only way to actually get a `ModulePath` is through
a `ModuleTracker` which is pretty intuitive and suggests that this is a good name.

This will free up the name `SourcePath` so that we can use it as a semantic
path type that wraps a raw `PyrePath.t`

Note: most of the code changes were autogenerated via
```
> rg 'SourcePath\.' --glob '!_build' -l | xargs xargs sed -i 's/SourcePath\./ModulePath./g'
```

Reviewed By: grievejia

Differential Revision: D36665606

fbshipit-source-id: f4413ae98336ddf55770e4763348ef5781278d28