Add a new semantic path type SourcePath
Summary:
In my opinion distingiushing SourcePath from PyrePath
doesn't provide as much value as distinguishing ArtifactPath
(see previous commit) because:
- the main type safety benfit is to remind us of the differene,
  and for that it suffices to tag just one of the types so that
  we can't accidentally forget to convert
- the only thing we ever really do is convert a PyrePath to a
  SourcePath and back again in a naive way; unlike ArtifactPath
  there's no such thing as a nontrivial mapping between raw
  PyrePath and SourcePath

Nontheless, I do think the intent is a bit clearer (and likely
the semantic types are more intuitive) if we have newtypes for
both SourcePath and ArtifactPath. That's what this commit provides.

Reviewed By: grievejia

Differential Revision: D36667759

fbshipit-source-id: 2d01279f87ddfb0aaaa66e7e261db907a515d823