Remove ~additional_preprocessing, it is dead code
Summary:
We no longer use the `additional_preprocessing` argument to
`AstEnvironment` in production - pysa currently relies on a global
mutation to set decorator inlining.

Let's get rid of it - it's dead code, and it's getting in the way
of me making refactors. If, in the future, we need to make preprocessing
pluggable explicitly then we should use flags in `Configuration.Analysis.t`
for that; today there's no need and we shouldn't complexify our APIs.

Reviewed By: pradeep90

Differential Revision: D36717710

fbshipit-source-id: 3e333b52e3f43856203dff496ce618cb5eff4667