Factor out missing-from-stub data
Summary:
UnannotatedGlobalEnvironment is probably the most complex of all
our environments from an implementation perspective (in terms of the
environment data itself; AttributeResolution has more complex
data manipulations but the cache structures are completely vanilla).

As a result, readability of the code matters and it's pretty distracting
to have low-level manipulation of ast data for missing python entities
sitting around in this module. So, I'm moving them to a new module
that I call `MissingFromStubs` (happy to rename if reviewers have better
ideas, I thought this name was evokative enough that no one is likely
to be confused).

Reviewed By: grievejia

Differential Revision: D36739178

fbshipit-source-id: 991f49e7673d02898f6298926ef7f56e3581a06d