Add 'name' clause to all Pysa DSL ModelQueries in fbsource
Summary: Currently, the `name` clause for Pysa DSL `ModelQuery`s is optional. We want to make this clause mandatory so that each ModelQuery will have a unique identifier that can be used to query them and log stats. To prepare for that change, I added `name` clauses to all existing `ModelQuery`s in `.pysa` files in the `fbsource` repository.

Reviewed By: alexkassil

Differential Revision: D36728220

fbshipit-source-id: 4f411a415e7a74b89e7f67674bac8873cc2947a9