Make update a one-shot operation
Summary:
Up until now, the mechanism for updating an environment has been to:
- call the build system to get an ArtifactPath.t list
- extract the module tracker from the environment stack
- call `ModuleTracker.update` and save the result to a variable
- pass that result to the environment stack update functions

This is silly, there is nothing we would ever do with a module tracker
update result - outside tests where we can always tweak the flow - other
than use it to update the environments. And the environments all contain
a module tracker inside them anyway.

So, let's just call update and have that do the module tracker stage; now
the flow looks like:
- call the build system to get an ArtifactPath.t list
- update the environment

**Why is this commit not smaller?***

This change winds up being a bigger refactor than it sounds like it ought to because
- we were tied to the confusing way of doing things in tests
- in order to log module tracker updates from the server top-level, I had
  to add those to the update result type (otherwise they would be hidden;
having it be a hidden implementation detail of the AstEnvironment / ModuleTracker interaction
  which would be nice but doesn't play well with logging)

Reviewed By: grievejia

Differential Revision: D36673248

fbshipit-source-id: ac25c8f8bd9ef8f8f29c08113fbaef34f13932e5