Simplify environment: kill UpdateResult.Make functor
Summary:
There is no reason to functorize UpdateResult.Make - it is only
ever used once, inside of EnvironmentTable.Make, so inlining it
is simpler and does not break any meaningful abstraction boundary.

Moreover, I want to fix our tests and this requires adding a
Testing module to the environment stack that will let us peek
into inner environments and update results for grey-box testing
(in particular of dependency trigger logic) - the current structure
is getting in my way but a simpler structure will work just fine.

Reviewed By: pradeep90

Differential Revision: D36781375

fbshipit-source-id: 95456d6a7de7a56c31909c372ccf9b9ca72dce4c