Clean up Test: Stop calling `create` in tests
Summary:
One of the major goals of my current test refactor is that I want to
avoid our tests forking off multiple environments; from the same
scratch project; doing this never makes sense but is very easy to
do accidentally using the existing APIs; it frequently causes bugs
in update tests because the resulting environments aren't really valid
from a dependency tracking point of view.

This commit updates many of the environment tests to use the standard
pattern of calling `global_environment` on `project` to get the
environment stack and then using `AnnotatedGlobalEnvironment.Testing`
functions to reach down to the layer we want to test.

Differential Revision: D36792514

fbshipit-source-id: 0d0e7dfda9e6d69d65982c92374042319cb16bd8