Clean up test: add Testing modules to get inner environments
Summary:
The most confusing thing about Test.ml is that there are far
too many codepaths to create environments. This is a challenge
for multiple reasons:
- it leads to code duplication
- there are many ways to create environments so our test code
  isn't as consistent as it could be with a cleaner API
- many ways of doing this are illegal (because it's illegal
  to nest the same module tracker in separate environments) but
  there's no way to enforce legality
- there's no good way to write an update interface that enforces
  certain arbitrary restrictions (e.g. in-memory projects cannot
  be updated) that we want to enforce.

NOTE: it would be nicer long term to have this nesting only for read-only environments; that doesn't work with the current structure of the test code. I make that change later on, in D36781373

Reviewed By: grievejia

Differential Revision: D36781376

fbshipit-source-id: 7d1e3e01afb9eedcbb14b817df83f68f82763f59