Move the caching logic under the taint analysis directory
Summary:
The pysa cache logic currently lives under `service/staticAnalysis.ml`. This file is meant to provide re-usable services to build static analyses.
In practice, that code is highly coupled with the taint analysis, since it checks for changes in `.pysa` files.
Let's move the logic under the taint analysis directory to clean this up.

Differential Revision: D36851663

fbshipit-source-id: d62ddad65bde8a22e77acfe03807f2d602579e73