Enable pruning of the dependency graph in integration tests
Summary:
During the taint analysis, after computing the dependency graph, we prune the graph by performing a depth first search on the reverse dependency graph (mapping of callers to callees), starting from all functions in the source directory to analyze. This allows us to skip the analysis of all functions that can never be reached, hence saving a lot of time.

This diff enables the pruning in our integration tests, for consistency with the actual analysis.

Differential Revision: D36851786

fbshipit-source-id: 89506957019cae901f1675132777d6cc818e1270