Clean up Test: only expose read-only environments
Summary:
This commit makes two lock-step changes:
- it updates the `Testing` logic in D36781376 (https://github.com/facebook/pyre-check/commit/c86d9ff838af0bb3c232223beea3f8a690071a55) so that we have convenient code
  to grab inner environments.
- it updates the test apis to expose read-only environments (in the process
  forcing all callers of the `update` logic to use
  `ScratchProject.update_global_environment`, which is needed to ensure we
  can't introduce certain kinds of hard-to-debug testing bugs

Something I had to do as part of this was expose the read-only module tracker
as part of the AstEnvironment.ReadOnly api; this is very standard (most of our
read-only environments expose their inner layers, which is always safe to do
as long as everything is read-only).

Unfortunately there are quite a lot of individual line changes required all in
one commit, but they are all just trivial consequences of switching to
read-only apis.

**Notes**

The only non-readonly layer that we directly expose to users is ModuleTracker,
which is necessary for grey-box testing because we can't see the inner
ModuleTracker update results from the current UpdateResult type; both
ModuleTrackerTest and ChangedPathsTest have to do this.

I removed the `TypeEnvironment.invalidate` logic from interprocedural tests;
this isn't needed now that type environment uses first-class tables and the
fact that it isn't needed is self-critiquing given that the tests pass.

Differential Revision: D36792515

fbshipit-source-id: 3d1c5e140fdefa8370bd8710bd1538c97cf9bd56