Add a handle to the call graph in shared memory
Summary:
Right now, the call graph in shared memory is accessible from any module, as a global value.
This leads to bugs where we access the call graph before it is initialized.
To prevent this, let's add a dummy type `t = Handle` which is returned by the call graph construction.

Differential Revision: D36885902

fbshipit-source-id: b88036d166c29b331245d7e36ce425c78e2d28f0