Rename `IntervalSet` into `ClassIntervalSet`
Summary: We have a module `ClassInterval` which represents a class interval, then we have a module `IntervalSet` which is a set of `ClassInterval`. This diff renames `IntervalSet` into `ClassIntervalSet` to prevent confusion. `IntervalSet` cannot be used to represent an arbitrary set of intervals, hence I believe we should keep the `Class` prefix.

Differential Revision: D36888408

fbshipit-source-id: b738a0be5b523a6e7cae8e7fd0ca7e29bd53cd8e