Get rid of `cold_start` - handle it in create
Summary:
Ever since we introduced lazy parsing, it is extremely cheap to
cold start an environment - on the order of a millisecond or so.

Moreover, having it as a separate function causes some issues:
- it's currently possible to try to create a read-only environment
  and use it without having cold-started yet, which would cause
  problems for lazy loading
- as soon as we add overlays, we'll have a new problem where trying
  to cold start an overlay doesn't make any sense, we only want to
  do that for brand-new environments

These problems all go away if cold-starting UnannotatedGlobalEnvironment
(which just involves loading `builtins` to make later lazy loads faster)
happens as part of create.

Differential Revision: D36794771

fbshipit-source-id: 3a9389d6956c70c6b326046f5c03f1c2c15650e8