Standardize TypeEnvironment [1] constructors
Summary:
This commit standardizes the functions to create and save a type
environment to be the same as for all other environments:
- create to make a new, production environment
- create_for_testing to make a test environment with in-memory sources
- store to create a saved state
- load to load from a saved state

Additionally, since this is the top level of the environment stack I
thought it was a reasonable place to put the shared memory keys loading
and saving, which is an important part of saved-state handling because
without it the saved states can't do incremental updates.

The use-case code gets simpler with this change because we no longer
manually insert or extract the global environment when performing these
operations.

And importantly, the implementations are now much more future
proof (remember that the implicit load and store will not work if we
switch to ocaml 5 and non-shared-memory tables; this new API on the other
hand will work fine)

Reviewed By: grievejia

Differential Revision: D36921756

fbshipit-source-id: f4ec87ed9caefd4fad1dc3aa00c0d4da019da505