create AstEnvironment.Base
Summary:
The next three commits present one possible way to make the distinction between base and overlay environments type-safe (which will also make overlays much easier to write, since it enforces code modularity that simplifies each step).

The idea is to create a new, non-public module that represents the logic to manage the cache of an environment. This layer is
- not readonly because it knows about the underlying table, and can both lazy-load and update or remove keys
- but isn't a full read-write environment because it has only `ReadOnly` access to lower layers

For now I call this type `Abstract`, and it generally looks just like the current environment type, except that the upstream is a read-only rather than read-write.

What's the benefit? There are three benefits:
- one is that the *existing* logic becomes more type-safe. Right now there's nothing in `update_only_this_environment` guaranteeing that we never make use of the non-read-only upstream, but by moving this code into `Abstract` we can trivially show at the type level that it always depends only on read-only access to upstreams
- the bigger benefit is that all this code is going to be shared by overlays and base environments, but *none of the other code will be*: the create, load, and update functions for base environments are completely non-reusable for overlays.
- finally, the resulting APIs after we add overlays will be type-safe because `Base.t` and `Overlay.t` will be two different types, with exactly the right functions defined on them; using a non-type-safe "flag" in a single type wouldn't work for this

To emphasize this, I also move those functions into a new module called `Base`, which I then include so as to keep the interfaces the same for now. When I implement overlays, I'll do it by creating a new module `Overlay` at the same level as `Base`, which will rely on `Abstract` and `ReadOnly` but will have completely different functions for creating, updating, and propagating parent updates.

NOTE: I'm not attached to the names `Abstract` and `Base`; I think `Base` is probably a decent name but `Abstract` should probably be something else, maybe `Cache` or `CacheLogic` or `CacheHandler`. Open to suggestions; FWIW this name will never leak into public interfaces so it's trivial to change later.

**How will `Overlay` actually work?**

The actual implementation of `Overlay` - which will be based on `ReadOnly` and `Abstract` but won't use any logic from `Base` - will involve doing ownership filters both on reads and updates.

Updates are pretty easy: we can adapt the existing triggered dependencies logic but add a filter for ownership, and then pass it to a function that updates just the necessary keys (the functionality for that update can be shared, and live in `Abstract`)

The read-only will work by creating two different read-only environments (the parent and the overlay) and then choosing which one to use for a specific read based on ownership checks.

**Could I do this whole thing by making Abstract a functor?**

Yes, we could make the upstream read-only something exposed via an input module and functorize everything.

But functors are terrible in 4.10.2 - by that I don't mean they are inherently terrible, but the compiler error messages are *hard* to read.

So, given that I have a non-functor-based solution that works about as well, I would much rather get it implemented using plain modules; we can always functorize after the fact but I *really, really* don't want to spend my overlay sprint puzzling over poorly-expressed functor type mismatch errors.

Reviewed By: grievejia

Differential Revision: D36902470

fbshipit-source-id: 55189e78de66c46d226feafbce719a5d8f69585a